更新Coremark

+++ bsp/lib/include/uart.h
@@ -18,7 +18,7 @@
 
 void uart_enable_ctr(uint32_t UARTx, uint32_t uart_en);//串口使能控制
 void uart_band_ctr(uint32_t UARTx, uint32_t uart_band);//串口波特率控制
-void uart_send_date(uint32_t UARTx, uint32_t uart_send);//串口发送
+void uart_send_date(uint32_t UARTx, uint8_t uart_send);//串口发送
 uint8_t uart_recv_date(uint32_t UARTx);//串口接收
 uint8_t uart_recv_flg(uint32_t UARTx);//串口接收状态查询
